+++ AnouQuickSeriesApp/app/src/main/res/values/strings.xml
@@ -2,5 +2,5 @@
     <string name="action_settings">Settings</string>
     <string name="section_format">Hello World from section: %1$d</string>
     <string name="tab_bar_restaurant">Restaurants</string>
-    <string name="tab_bar_vacation">Vacation spots</string>
+    <string name="tab_bar_vacation">Vacation Spots</string>
 </resources>
