--- submodules/quickseries-core/src/main/java/com/quickseries/data/Restaurant.java
@@ -1,18 +1,0 @@
-package com.quickseries.data;
-
-/**
- * Created by Anou on 2017-10-14.
- */
-
-public class Restaurant {
-
-    private final String title;
-
-    public Restaurant(String title) {
-        this.title = title;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-}
