--- submodules/quickseries-core/src/main/java/com/quickseries/data/Vacation.java
@@ -1,8 +1,0 @@
-package com.quickseries.data;
-
-/**
- * Created by Anou on 2017-10-14.
- */
-
-public class Vacation {
-}
